+++ PacletInfo.m
@@ -1,18 +1,19 @@
 (* Paclet Info File *)
 
-(* created 2019/01/16*)
+(* created 2019/03/16*)
 
 Paclet[
     Name -> "MaXrd",
-    Version -> "1.1.0",
+    Version -> "1.2.0",
     MathematicaVersion -> "11.3+",
     Description -> "Database and utilities related to crystallography and X-ray scattering.",
     Creator -> "Stian Penev Ramsnes",
+    URL -> "https://github.com/Stianpr20/MaXrd",
     Extensions -> 
         {
-            {"Documentation", Resources -> 
-                {"Guides/MaXrd"}
-            , Language -> "English", MainPage -> "Guides/MaXrd"}
+            {"Application", Context -> "MaXrd`"}, 
+            {"Documentation", Language -> "English", MainPage -> "Guides/MaXrd"}, 
+            {"Documentation", Resources -> {"Guides/MaXrd"} }
         }
 ]